refactor: move InputTokenValue to PoolOperations
