+++ src/components/PoolOperations/InputTokenValue/index.tsx
@@ -6,9 +6,9 @@ import Big from 'big.js'
 import Tippy from '@tippyjs/react'
 import 'tippy.js/dist/tippy.css'
 
-import { usePoolTokens } from '../../context/PoolTokensContext'
-import { BNtoDecimal } from '../../utils/numerals'
-import { priceDollar } from '../../utils/priceDollar'
+import { usePoolTokens } from '../../../context/PoolTokensContext'
+import { BNtoDecimal } from '../../../utils/numerals'
+import { priceDollar } from '../../../utils/priceDollar'
 import { Input } from './styles'
 
 interface IInputProps {
