+++ src/components/PoolOperations/InputTokenValue/styles.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import theme from '../../styles/theme'
+import theme from '../../../styles/theme'
 
 export const Input = styled.input`
   background-color: transparent;
